fix: Skip MAX_ACC_OUT_READ on the broken FW versions
According to the PRM, setting MAX_ACC_OUT_READ to zero enables the auto mode,
which applies the best suitable optimizations.
However, there is a bug in certain FW versions, where the zero value is not available.
In this case, until the fix is available, skipping this parameter and emitting a warning

Signed-off-by: Alexander Maslennikov <amaslennikov@nvidia.com>