switch from mstconfig tool to mlxconfig
mstconfig is a version of the mlxconfig tool,
distributed via the package manager. That means it only receives
the necessary security fixes and might not feature the up-to-date
functionality. Instead of that, let's switch to using the mlxconfig,
which is distributed as part of the MLNX_OFED packages. That way,
we are always using the newest version.

In the future, we can add a CI job that would periodically query
the latest version from the ofed repo and update the tools versions
accordingly.

Signed-off-by: Alexander Maslennikov <amaslennikov@nvidia.com>